+++ deployment/puppet/inifile/spec/spec_helper_acceptance.rb
@@ -2,11 +2,16 @@ require 'beaker-rspec/spec_helper'
 require 'beaker-rspec/helpers/serverspec'
 
 unless ENV['RS_PROVISION'] == 'no'
-  if hosts.first.is_pe?
-    install_pe
+  # This will install the latest available package on el and deb based
+  # systems fail on windows and osx, and install via gem on other *nixes
+  foss_opts = {:default_action => 'gem_install'}
+
+  if default.is_pe?; then
+    install_pe;
   else
-    install_puppet
+    install_puppet(foss_opts);
   end
+
   hosts.each do |host|
     if host['platform'] =~ /debian/
       on host, 'echo \'export PATH=/var/lib/gems/1.8/bin/:${PATH}\' >> ~/.bashrc'
@@ -26,14 +31,7 @@ RSpec.configure do |c|
   c.before :suite do
     # Install module and dependencies
     hosts.each do |host|
-      if host['platform'] !~ /windows/i
-        copy_root_module_to(host, :source => proj_root, :module_name => 'inifile')
-      end
-    end
-    hosts.each do |host|
-      if host['platform'] =~ /windows/i
-        on host, puppet('plugin download')
-      end
+      copy_root_module_to(host, :source => proj_root, :module_name => 'inifile')
     end
   end