+++ deployment/puppet/inifile/spec/unit/puppet/provider/ini_setting/ruby_spec.rb
@@ -137,6 +137,9 @@ url = http://192.168.1.1:8080
 subby=bar
     #another comment
  ; yet another comment
+
+-nonstandard-
+  shoes = purple
       EOS
     }
 
@@ -164,6 +167,42 @@ yahoo = yippee
 subby=bar
     #another comment
  ; yet another comment
+
+-nonstandard-
+  shoes = purple
+      EOS
+)
+    end
+
+    it "should add a missing setting to the correct section with pre/suffix" do
+      resource = Puppet::Type::Ini_setting.new(common_params.merge(
+          :section => 'nonstandard',
+          :setting => 'yahoo', :value => 'yippee',
+          :section_prefix => '-', :section_suffix => '-'))
+      provider = described_class.new(resource)
+      provider.exists?.should be false
+      provider.create
+      validate_file(<<-EOS
+# This is a comment
+[section1]
+; This is also a comment
+foo=foovalue
+
+bar = barvalue
+master = true
+[section2]
+
+foo= foovalue2
+baz=bazvalue
+url = http://192.168.1.1:8080
+[section:sub]
+subby=bar
+    #another comment
+ ; yet another comment
+
+-nonstandard-
+  shoes = purple
+  yahoo = yippee
       EOS
 )
     end
@@ -191,6 +230,9 @@ url = http://192.168.1.1:8080
 subby=bar
     #another comment
  ; yet another comment
+
+-nonstandard-
+  shoes = purple
 yahoo = yippee
       EOS
 )
@@ -219,6 +261,41 @@ url = http://192.168.1.1:8080
 subby=bar
     #another comment
  ; yet another comment
+
+-nonstandard-
+  shoes = purple
+      EOS
+      )
+    end
+
+    it "should modify an existing setting with pre/suffix with a different value" do
+      resource = Puppet::Type::Ini_setting.new(common_params.merge(
+           :section => 'nonstandard',
+           :setting => 'shoes', :value => 'orange',
+           :section_prefix => '-', :section_suffix => '-' ))
+      provider = described_class.new(resource)
+      provider.exists?.should be true
+      provider.value=('orange')
+      validate_file(<<-EOS
+# This is a comment
+[section1]
+; This is also a comment
+foo=foovalue
+
+bar = barvalue
+master = true
+[section2]
+
+foo= foovalue2
+baz=bazvalue
+url = http://192.168.1.1:8080
+[section:sub]
+subby=bar
+    #another comment
+ ; yet another comment
+
+-nonstandard-
+  shoes = orange
       EOS
       )
     end
@@ -247,6 +324,9 @@ url = http://192.168.1.1:8080
 subby=foo
     #another comment
  ; yet another comment
+
+-nonstandard-
+  shoes = purple
       EOS
       )
     end
@@ -276,6 +356,43 @@ url = http://192.168.0.1:8080
 subby=bar
     #another comment
  ; yet another comment
+
+-nonstandard-
+  shoes = purple
+    EOS
+      )
+    end
+
+    it "should be able to handle settings with pre/suffix with non alphanumbering settings " do
+      resource = Puppet::Type::Ini_setting.new(common_params.merge(
+           :section => 'nonstandard',
+           :setting => 'shoes', :value => 'http://192.168.0.1:8080',
+           :section_prefix => '-', :section_suffix => '-' ))
+      provider = described_class.new(resource)
+      provider.exists?.should be true
+      provider.value.should == 'purple'
+      provider.value=('http://192.168.0.1:8080')
+
+      validate_file( <<-EOS
+# This is a comment
+[section1]
+; This is also a comment
+foo=foovalue
+
+bar = barvalue
+master = true
+[section2]
+
+foo= foovalue2
+baz=bazvalue
+url = http://192.168.1.1:8080
+[section:sub]
+subby=bar
+    #another comment
+ ; yet another comment
+
+-nonstandard-
+  shoes = http://192.168.0.1:8080
     EOS
       )
     end
@@ -283,6 +400,15 @@ subby=bar
     it "should recognize an existing setting with the specified value" do
       resource = Puppet::Type::Ini_setting.new(common_params.merge(
            :setting => 'baz', :value => 'bazvalue'))
+      provider = described_class.new(resource)
+      provider.exists?.should be true
+    end
+
+    it "should recognize an existing setting with pre/suffix with the specified value" do
+      resource = Puppet::Type::Ini_setting.new(common_params.merge(
+           :section => 'nonstandard',
+           :setting => 'shoes', :value => 'purple',
+           :section_prefix => '-', :section_suffix => '-' ))
       provider = described_class.new(resource)
       provider.exists?.should be true
     end
@@ -311,7 +437,44 @@ subby=bar
     #another comment
  ; yet another comment
 
+-nonstandard-
+  shoes = purple
+
 [section3]
+huzzah = shazaam
+      EOS
+      )
+    end
+
+    it "should add a new section with pre/suffix if the section does not exist" do
+      resource = Puppet::Type::Ini_setting.new(common_params.merge(
+          :section => "section3", :setting => 'huzzah', :value => 'shazaam',
+          :section_prefix => '-', :section_suffix => '-' ))
+      provider = described_class.new(resource)
+      provider.exists?.should be false
+      provider.create
+      validate_file(<<-EOS
+# This is a comment
+[section1]
+; This is also a comment
+foo=foovalue
+
+bar = barvalue
+master = true
+[section2]
+
+foo= foovalue2
+baz=bazvalue
+url = http://192.168.1.1:8080
+[section:sub]
+subby=bar
+    #another comment
+ ; yet another comment
+
+-nonstandard-
+  shoes = purple
+
+-section3-
 huzzah = shazaam
       EOS
       )
@@ -341,7 +504,44 @@ subby=bar
     #another comment
  ; yet another comment
 
+-nonstandard-
+  shoes = purple
+
 [section:subsection]
+huzzah = shazaam
+      EOS
+      )
+    end
+
+    it "should add a new section with pre/suffix if the section does not exist - with colon" do
+      resource = Puppet::Type::Ini_setting.new(common_params.merge(
+          :section => "section:subsection", :setting => 'huzzah', :value => 'shazaam',
+          :section_prefix => '-', :section_suffix => '-' ))
+      provider = described_class.new(resource)
+      provider.exists?.should be false
+      provider.create
+      validate_file(<<-EOS
+# This is a comment
+[section1]
+; This is also a comment
+foo=foovalue
+
+bar = barvalue
+master = true
+[section2]
+
+foo= foovalue2
+baz=bazvalue
+url = http://192.168.1.1:8080
+[section:sub]
+subby=bar
+    #another comment
+ ; yet another comment
+
+-nonstandard-
+  shoes = purple
+
+-section:subsection-
 huzzah = shazaam
       EOS
       )
@@ -359,6 +559,19 @@ setting1 = hellowworld
 ", emptyfile)
     end
 
+    it "should add a new section with pre/suffix if no sections exists" do
+      resource = Puppet::Type::Ini_setting.new(common_params.merge(
+          :section => "section1", :setting => 'setting1', :value => 'hellowworld', :path => emptyfile,
+          :section_prefix => '-', :section_suffix => '-' ))
+      provider = described_class.new(resource)
+      provider.exists?.should be false
+      provider.create
+      validate_file("
+-section1-
+setting1 = hellowworld
+", emptyfile)
+    end
+
     it "should add a new section with colon if no sections exists" do
       resource = Puppet::Type::Ini_setting.new(common_params.merge(
           :section => "section:subsection", :setting => 'setting1', :value => 'hellowworld', :path => emptyfile))
@@ -371,6 +584,18 @@ setting1 = hellowworld
 ", emptyfile)
     end
 
+    it "should add a new section with pre/suffix with colon if no sections exists" do
+      resource = Puppet::Type::Ini_setting.new(common_params.merge(
+          :section => "section:subsection", :setting => 'setting1', :value => 'hellowworld', :path => emptyfile,
+          :section_prefix => '-', :section_suffix => '-' ))
+      provider = described_class.new(resource)
+      provider.exists?.should be false
+      provider.create
+      validate_file("
+-section:subsection-
+setting1 = hellowworld
+", emptyfile)
+    end
     it "should be able to handle variables of any type" do
       resource = Puppet::Type::Ini_setting.new(common_params.merge(
           :section => "section1", :setting => 'master', :value => true))
@@ -495,26 +720,6 @@ foo=bar
       EOS
     }
 
-    it "should fail if the separator doesn't include an equals sign" do
-      expect {
-        Puppet::Type::Ini_setting.new(common_params.merge(
-                                         :section           => 'section2',
-                                         :setting           => 'foo',
-                                         :value             => 'yippee',
-                                         :key_val_separator => '+'))
-      }.to raise_error Puppet::Error, /must contain exactly one/
-    end
-
-    it "should fail if the separator includes more than one equals sign" do
-      expect {
-        Puppet::Type::Ini_setting.new(common_params.merge(
-                                         :section           => 'section2',
-                                         :setting           => 'foo',
-                                         :value             => 'yippee',
-                                         :key_val_separator => ' = foo = '))
-      }.to raise_error Puppet::Error, /must contain exactly one/
-    end
-
     it "should modify an existing setting" do
       resource = Puppet::Type::Ini_setting.new(common_params.merge(
                                                    :section           => 'section2',
@@ -532,19 +737,46 @@ foo=yippee
       )
     end
 
+  end
+
+  context "when overriding the separator to something other than =" do
+    let(:orig_content) {
+      <<-EOS
+[section2]
+foo: bar
+      EOS
+    }
+
+    it "should modify an existing setting" do
+      resource = Puppet::Type::Ini_setting.new(common_params.merge(
+                                                   :section           => 'section2',
+                                                   :setting           => 'foo',
+                                                   :value             => 'yippee',
+                                                   :key_val_separator => ': '))
+      provider = described_class.new(resource)
+      provider.exists?.should be true
+      provider.value.should == 'bar'
+      provider.value=('yippee')
+      validate_file(<<-EOS
+[section2]
+foo: yippee
+      EOS
+      )
+    end
+
     it "should add a new setting" do
       resource = Puppet::Type::Ini_setting.new(common_params.merge(
                                                    :section           => 'section2',
                                                    :setting           => 'bar',
                                                    :value             => 'baz',
-                                                   :key_val_separator => '='))
+                                                   :key_val_separator => ': '))
       provider = described_class.new(resource)
       provider.exists?.should be false
       provider.create
       validate_file(<<-EOS
 [section2]
-foo=bar
-bar=baz
+foo: bar
+bar: baz
       EOS
       )
     end
@@ -569,6 +801,9 @@ url = http://192.168.1.1:8080
 subby=bar
     #another comment
  ; yet another comment
+
+ -nonstandard-
+   shoes = purple
 EOS
     }
 
@@ -593,6 +828,38 @@ url = http://192.168.1.1:8080
 subby=bar
     #another comment
  ; yet another comment
+
+ -nonstandard-
+   shoes = purple
+EOS
+    )
+    end
+
+    it "should remove a setting with pre/suffix that exists" do
+      resource = Puppet::Type::Ini_setting.new(common_params.merge(
+      :section => 'nonstandard', :setting => 'shoes', :ensure => 'absent',
+      :section_prefix => '-', :section_suffix => '-' ))
+      provider = described_class.new(resource)
+      provider.exists?.should be true
+      provider.destroy
+      validate_file(<<-EOS
+[section1]
+; This is also a comment
+foo=foovalue
+
+bar = barvalue
+master = true
+[section2]
+
+foo= foovalue2
+baz=bazvalue
+url = http://192.168.1.1:8080
+[section:sub]
+subby=bar
+    #another comment
+ ; yet another comment
+
+ -nonstandard-
 EOS
     )
     end
@@ -619,11 +886,43 @@ url = http://192.168.1.1:8080
 subby=bar
     #another comment
  ; yet another comment
+
+ -nonstandard-
+   shoes = purple
+      EOS
+      )
+    end
+
+    it "should do nothing for a setting with pre/suffix that does not exist" do
+      resource = Puppet::Type::Ini_setting.new(common_params.merge(
+         :section => 'nonstandard', :setting => 'foo', :ensure => 'absent',
+         :section_prefix => '-', :section_suffix => '-' ))
+      provider = described_class.new(resource)
+      provider.exists?.should be false
+      provider.destroy
+      validate_file(<<-EOS
+[section1]
+; This is also a comment
+foo=foovalue
+
+bar = barvalue
+master = true
+[section2]
+
+foo= foovalue2
+baz=bazvalue
+url = http://192.168.1.1:8080
+[section:sub]
+subby=bar
+    #another comment
+ ; yet another comment
+
+ -nonstandard-
+   shoes = purple
       EOS
       )
     end
   end
-
 
   context "when dealing with indentation in sections" do
     let(:orig_content) {